requirements.txt
ai2thor==5.0.0
ai2thor-colab==0.1.2
annotated-types==0.7.0
anyio==4.11.0
appnope==0.1.4
asttokens==3.0.0
aws-requests-auth==0.4.3
blinker==1.9.0
botocore==1.40.54
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
comm==0.2.3
contourpy==1.3.0
cryptography==46.0.3
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
einops==0.8.1
exceptiongroup==1.3.0
executing==2.2.1
faiss-cpu==1.12.0
filelock==3.19.1
Flask==3.1.2
fonttools==4.60.1
fsspec==2025.9.0
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==1.0.1
idna==3.11
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipykernel==6.30.1
ipython==8.18.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.7
MarkupSafe==3.0.3
matplotlib==3.9.4
matplotlib-inline==0.1.7
moviepy==2.2.1
mpmath==1.3.0
msgpack==1.1.2
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.1
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
prior==1.0.3
proglog==0.1.12
progressbar2==4.5.0
prompt_toolkit==3.0.52
protobuf==6.33.1
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
PyGithub==2.8.1
Pygments==2.19.2
PyJWT==2.10.1
PyNaCl==1.6.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-utils==3.9.1
python-xlib==0.33
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
regex==2025.11.3
requests==2.32.5
safetensors==0.6.2
scipy==1.13.1
sentry-sdk==2.45.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
stack-data==0.6.3
sympy==1.14.0
torch==2.3.1
torchinfo==1.8.0
torchvision==0.18.1
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==1.26.20
wandb==0.23.0
wcwidth==0.2.14
Werkzeug==3.1.3
zipp==3.23.0
